fix doctests

+++ qmcpy/integrand/barrier_option.py
@@ -21,7 +21,7 @@ class BarrierOption(Integrand):
         dim_frac        0
     >>> x = barrier_option.discrete_distrib.gen_samples(2**12)
     >>> y = barrier_option.f(x)
-    >>> y.mean()
+    >>> float(y.mean())
     1.090...
     >>> level_dims = [2,4,8]
     >>> barrier_option_multilevel = BarrierOption(DigitalNetB2(seed=7),multilevel_dims=level_dims)
@@ -32,7 +32,7 @@ class BarrierOption(Integrand):
     ...     x = barrier_option_single_level.discrete_distrib.gen_samples(2**12)
     ...     level_est = barrier_option_single_level.f(x).mean()
     ...     yml += level_est
-    >>> yml
+    >>> float(yml)
     1.103...
     """
